Fix db_client schema checks
